+++ tests/testthat/test01_auth.R
@@ -22,38 +22,11 @@ test_that("Authentication works",
     expect_is(az2, "az_rm")
     expect_true(is_azure_token(az2$token))
 
-    token <- az2$token
-
-    az3 <- az_rm$new(token=token)
+    az3 <- get_azure_login(tenant=tenant, app=app, password=password)
     expect_is(az3, "az_rm")
-    expect_true(is_azure_token(az3$token))
+
+    az4 <- get_azure_login(config_file=creds)
+    expect_is(az4, "az_rm")
 })
 
 
-test_that("Persistent authentication works",
-{
-    expect_true(suppressWarnings(is.null(delete_azure_login(tenant, confirm=FALSE))))
-    expect_true(all(names(list_azure_logins()) != tenant))
-
-    login_dirs <- rappdirs::user_config_dir("AzureRMR", "AzureR", roaming=FALSE)
-    expect_true(all(dir(login_dirs) != tenant))
-
-    az <- create_azure_login(tenant=tenant, app=app, password=password)
-    expect_is(az, "az_rm")
-    expect_true(length(dir(login_dirs, pattern=tenant)) == 1)
-
-    expect_true(is.null(delete_azure_login(tenant, confirm=FALSE)))
-    expect_true(all(names(list_azure_logins()) != tenant))
-    
-    creds <- tempfile(fileext=".json")
-    writeLines(jsonlite::toJSON(list(tenant=tenant, app=app, password=password)), creds)
-
-    az2 <- create_azure_login(tenant=tenant, app=app, password=password)
-    expect_is(az2, "az_rm")
-    expect_true(length(dir(login_dirs, pattern=tenant)) == 1)
-
-    expect_true(az$tenant == az2$tenant && az$tenant == tenant)
-
-    az3 <- get_azure_login(tenant)
-    expect_true(az$tenant == az3$tenant)
-})