+++ NEWS.md
@@ -2,12 +2,14 @@
 
 ## Significant interface changes
 
-* New `create_azure_login`, `get_azure_login` and `delete_azure_login` functions to handle ARM authentication. These will persist the login object across sessions, removing the need to re-authenticate each time. While directly calling `az_rm$new()` will still work, it's recommended to use `create_azure_login` and `get_azure_login` going forward.
-* `get_azure_token` revamped, now supports four authentication methods for obtaining AAD tokens:
+* New `get_azure_login` function to handle ARM authentication. While directly calling `az_rm$new()` will still work, it's recommended to use `get_azure_login` going forward. Login credentials will be saved and reused for subsequent sessions (see below).
+* `get_azure_token` significantly revamped. It now supports four authentication methods for obtaining AAD tokens:
   - Client credentials (what you would use with a "web app" registered service principal)
   - Authorization code (for a "native" service principal)
   - Device code
   - With a username and password (resource owner grant)
+* `get_azure_token` will now cache AAD tokens and refresh them for subsequent sessions. Tokens are cached in a user-specific configuration directory (unlike httr, which saves them in a special file in the R working directory).
+* Token acquisition logic will shortly move to a new package, to allow it to be used by other packages independently of the Resource Manager interface.
 
 ## Other changes